Refactor chat logic into useChat hook and ChatInterface component
- Create src/hooks/useChat.ts to manage chat state and lifecycle
- Create src/app/components/ChatInterface.tsx for pure UI rendering
- Update src/app/page.tsx to use the new hook and component
- Add src/types/chat.ts for Message interface
- Move messages, typing state, input handling, and allocation parsing into hook
