+++ frontend/src/app/page.tsx
@@ -1,24 +1,15 @@
 "use client"
-import React, { useState, useEffect, useRef } from 'react';
-import { Bot, Send, Target, Activity, CheckCircle2, AlertCircle } from 'lucide-react';
+import React, { useState } from 'react';
+import { Target, Activity } from 'lucide-react';
 import { PortfolioStats } from './components/PortfolioStats';
-import { evaluateGoalStatus, formatCurrency, getStatusColor, type GoalData } from '../utils/goalProjection';
+import { evaluateGoalStatus, getStatusColor, type GoalData } from '../utils/goalProjection';
 import PortfolioChart from './PortfolioChart';
-import { parseAllocationsFromMessage, getDefaultAllocations } from '../utils/allocationParser';
-import type { AssetAllocation } from '../utils/chartUtils';
-import { useNotifications } from '../hooks/useNotifications';
 import { DashboardHeader } from './components/DashboardHeader';
 import { ConnectWalletButton } from './components/ConnectWalletButton';
 import { useWallet } from './components/WalletContext';
 import { ErrorBoundary } from './components/ErrorBoundary';
-
-interface Message {
-  id: number;
-  sender: 'agent' | 'user';
-  text: string;
-  proactive?: boolean;
-  timestamp?: string;
-}
+import { useChat } from '../hooks/useChat';
+import { ChatInterface } from './components/ChatInterface';
 
 export default function Home() {
   const { publicKey, setPublicKey } = useWallet();
@@ -32,21 +23,10 @@ export default function Home() {
       try {
         const key = await window.freighterApi.getPublicKey();
         setPublicKey(key);
-      } catch (e) {
+      } catch {
         // Optionally handle rejection
       }
     };
-  const [messages, setMessages] = useState<Message[]>([
-    { id: 1, sender: 'agent', text: "Welcome to Smasage! 👋 I'm OpenClaw, your personal AI savings assistant natively built on Stellar. What financial goal can we crush today?" }
-  ]);
-  const [inputState, setInputState] = useState('');
-  const [isTyping, setIsTyping] = useState(false);
-
-  const [progress, setProgress] = useState(0);
-  const [goalStatus, setGoalStatus] = useState<'On Track' | 'Ahead' | 'Falling Behind'>('On Track');
-  const [allocations, setAllocations] = useState<AssetAllocation[]>(getDefaultAllocations());
-  const [wsConnected, setWsConnected] = useState(false);
-  const messagesEndRef = useRef<HTMLDivElement>(null);
 
   // Goal data
   const goalData: GoalData = {
@@ -57,93 +37,20 @@ export default function Home() {
     expectedAPY: 8.5,
   };
 
-  // WebSocket notifications
-  const { registerGoal } = useNotifications({
+  // Calculate goal status
+  const goalResult = evaluateGoalStatus(goalData);
+  const progress = goalResult.progressPercentage;
+  const goalStatus = goalResult.status;
+
+  const { messages, inputState, setInputState, isTyping, handleSendMessage, allocations, wsConnected } = useChat({
     userId: 'user-demo-001',
-    onNotification: (notification) => {
-      if (notification.type === 'connected') {
-        console.log('[App] Connected to notification server');
-        setWsConnected(true);
-      } else if (notification.type === 'agent-message') {
-        const payload = notification.payload as any;
-        const agentMsg: Message = {
-          id: Date.now(),
-          sender: 'agent',
-          text: payload.text,
-          proactive: payload.proactive,
-          timestamp: payload.timestamp,
-        };
-        setMessages(prev => [...prev, agentMsg]);
-        
-        // Parse allocations if present
-        const parsedAllocations = parseAllocationsFromMessage(payload.text);
-        if (parsedAllocations) {
-          setAllocations(parsedAllocations);
-        }
-      }
-    },
-    onError: (error) => {
-      console.error('[App] WebSocket error:', error);
-    },
-    enabled: true,
+    goalData,
   });
-
-  // Auto scroll
-  useEffect(() => {
-    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
-  }, [messages, isTyping]);
-
-  // Calculate goal status dynamically
-  useEffect(() => {
-    const result = evaluateGoalStatus(goalData);
-    setGoalStatus(result.status);
-    setProgress(result.progressPercentage);
-  }, []);
-
-  // Register goal with notification server on mount
-  useEffect(() => {
-    if (wsConnected) {
-      registerGoal({
-        currentBalance: goalData.currentBalance,
-        targetAmount: goalData.targetAmount,
-        targetDate: goalData.targetDate,
-        expectedAPY: goalData.expectedAPY,
-        monthlyContribution: goalData.monthlyContribution,
-      });
-    }
-  }, [wsConnected]);
-
-  const handleSendMessage = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!inputState.trim()) return;
-
-    const userMsg: Message = { id: Date.now(), sender: 'user', text: inputState };
-    setMessages(prev => [...prev, userMsg]);
-    setInputState('');
-    setIsTyping(true);
-
-    // Mock agent response delay
-    setTimeout(() => {
-      setIsTyping(false);
-      const agentMsg: Message = {
-        id: Date.now() + 1,
-        sender: 'agent',
-        text: "That's a great goal. I'll allocate 60% to Stellar Blend for safe yield, and the rest to Soroswap XLM/USDC LP to accelerate returns. Does that sound good?"
-      };
-      setMessages(prev => [...prev, agentMsg]);
-      
-      // Parse allocations from agent message
-      const parsedAllocations = parseAllocationsFromMessage(agentMsg.text);
-      if (parsedAllocations) {
-        setAllocations(parsedAllocations);
-      }
-    }, 1800);
-  };
 
   return (
     <ErrorBoundary fallbackMessage="The dashboard failed to load. Please try again.">
       <>
-        <DashboardHeader wsConnected={wsConnected}>
+      <DashboardHeader wsConnected={wsConnected}>
           <ConnectWalletButton onClick={handleConnectWallet} publicKey={publicKey || undefined} />
         </DashboardHeader>
         {showInstallModal && (
@@ -207,54 +114,13 @@ export default function Home() {
         </div>
 
         {/* Right Panel - Chat Agent */}
-        <div className="glass-panel">
-          <div className="chat-container">
-            <div className="chat-header">
-              <div className="agent-avatar">
-                <Bot size={28} />
-              </div>
-              <div>
-                <h2 style={{ margin: 0, fontSize: '1.25rem' }}>OpenClaw Agent</h2>
-                <div style={{ display: 'flex', alignItems: 'center', gap: '6px', fontSize: '0.85rem', color: 'var(--success)' }}>
-                  <CheckCircle2 size={12} fill="var(--success)" color="var(--bg-card)" /> Online
-                </div>
-              </div>
-            </div>
-
-            <div className="chat-messages">
-              {messages.map((msg) => (
-                <div key={msg.id} className={`message ${msg.sender}`}>
-                  {msg.proactive && (
-                    <div style={{ display: 'flex', alignItems: 'center', gap: '6px', fontSize: '0.75rem', color: 'var(--accent-primary)', marginBottom: '4px', fontWeight: 600, textTransform: 'uppercase' }}>
-                      <AlertCircle size={12} /> Proactive Nudge
-                    </div>
-                  )}
-                  <div className="message-bubble">{msg.text}</div>
-                </div>
-              ))}
-              {isTyping && (
-                <div className="message agent">
-                  <div className="typing-indicator">
-                    <span></span><span></span><span></span>
-                  </div>
-                </div>
-              )}
-              <div ref={messagesEndRef} />
-            </div>
-
-            <form onSubmit={handleSendMessage} className="chat-input-container">
-              <input
-                type="text"
-                placeholder="Ask Smasage to adjust goals..."
-                value={inputState}
-                onChange={(e) => setInputState(e.target.value)}
-              />
-              <button type="submit" className="send-button">
-                <Send size={18} />
-              </button>
-            </form>
-          </div>
-        </div>
+        <ChatInterface
+          messages={messages}
+          inputState={inputState}
+          setInputState={setInputState}
+          isTyping={isTyping}
+          handleSendMessage={handleSendMessage}
+        />
       </main>
       </>
     </ErrorBoundary>
