+++ frontend/src/types/chat.ts
@@ -1,0 +1,7 @@
+export interface Message {
+  id: number;
+  sender: 'agent' | 'user';
+  text: string;
+  proactive?: boolean;
+  timestamp?: string;
+}
